type correction, test simplification

+++ src/main/java/ch/uzh/ifi/hase/soprafs22/rest/dto/UnitHealthDTO.java
@@ -13,7 +13,7 @@ public class UnitHealthDTO {
         return unitPosition;
     }
 
-    public long getHealth() {
+    public int getHealth() {
         return health;
     }
 }
